Node - add supertest tests for Express apps, refactor express app to be more testable
