+++ server/api/tic-tac-toe/tic-tac-toe.js
@@ -7,10 +7,4 @@ const app = express();
 
 app.get('/new', (req, res) => res.json(new TicTacToe()));
 
-// app.post('/move', (req, res) => {
-//   const { board, player } = req.body;
-//   console.log('/move Board:', board);
-//   console.log('/move Player:', player);
-// });
-
 module.exports = app;
