+++ server/server.js
@@ -1,22 +1,8 @@
 
-const express = require('express');
-const path = require('path');
-const api = require('./api');
+const { app } = require('./app');
 
 const PORT = process.env.PORT || 8080;
-
-const app = express();
-app.use(express.static(path.join(__dirname, '../build')));
-app.use(express.json());
-
-// Add APIs to the app.
-api(app);
-app.get('/', (req, res) => {
-  res.sendFile(path.join(__dirname, '../build', 'index.html'));
-});
-
 app.listen(PORT);
 
-// TODO: Use some debug library.
 // eslint-disable-next-line no-console
 console.log(`Serving on port ${PORT}...`);
